% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTF.tool.R
\name{add_CDS}
\alias{add_CDS}
\title{Add CDS (Coding Sequence) Annotations to Genomic Data}
\usage{
add_CDS(input, genetic_elements = c("TRANSCRIPT", "MRNA", "CDS"))
}
\arguments{
\item{input}{A data frame containing genomic data. The data frame should have the following columns:
- `chr`: Chromosome identifier
- `start`: Start position of the annotation
- `end`: End position of the annotation
- `strand`: Strand information ('+' or '-')
- `annotationType`: Type of annotation (e.g., 'EXON', 'CDS')
- `gene_name`: Name of the associated gene}

\item{genetic_elements}{A character vector of annotation types to consider when checking for existing elements
(default: c("CDS", "TRANSCRIPT", "MRNA")). This is used to determine whether a given element type
already exists for a transcript; if it does, the CDS will not be created for that transcript.}
}
\value{
A data frame with the original input data and additional rows for the CDS.
  Each added row includes the following fields:
  - `source`: "JBIO-predicted" for newly added annotations
  - `annotationType`: Indicates 'CDS'
  - `start` and `end`: Updated start and end positions for the CDS
  - `strand`: Strand information copied from the input data
  - Other fields as present in the input data
}
\description{
This function generates coding sequence (CDS) regions for each transcript based on its associated exons.
It assumes that input data includes properly assigned exons and their transcript IDs. The output consists
of CDS regions suitable for further analysis or export in GTF/GFF format.
}
\details{
The function iterates over unique chromosomes and strand orientations, calculating CDS
for each gene.
}
\examples{

# Run the function
output_data <- add_CDS(input, genetic_elements = c("CDS", 'TRANSCRIPT', 'MRNA'))

}
